Add prerender pipeline for public routes
