+++ client/index.html
@@ -34,9 +34,10 @@
       content="SimReady scenes for robotics. OpenUSD, clean pivots and physics, Isaac-ready joints and colliders. Prove policies faster with realistic environments."
     />
     <meta name="robots" content="index, follow" />
+    <!--app-head-->
   </head>
   <body class="font-sans">
-    <div id="root"></div>
+    <div id="root"><!--app-html--></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
