+++ scripts/public-routes.ts
@@ -1,0 +1,18 @@
+export const publicRoutes = [
+  { path: "/", changefreq: "weekly", priority: 1.0 },
+  { path: "/why-simulation", changefreq: "monthly", priority: 0.8 },
+  { path: "/marketplace", changefreq: "daily", priority: 0.9 },
+  { path: "/solutions", changefreq: "monthly", priority: 0.8 },
+  { path: "/pricing", changefreq: "monthly", priority: 0.8 },
+  { path: "/learn", changefreq: "weekly", priority: 0.7 },
+  { path: "/docs", changefreq: "monthly", priority: 0.8 },
+  { path: "/evals", changefreq: "monthly", priority: 0.7 },
+  { path: "/benchmarks", changefreq: "monthly", priority: 0.7 },
+  { path: "/rl-training", changefreq: "monthly", priority: 0.7 },
+  { path: "/case-studies", changefreq: "monthly", priority: 0.7 },
+  { path: "/careers", changefreq: "weekly", priority: 0.6 },
+  { path: "/contact", changefreq: "monthly", priority: 0.6 },
+  { path: "/partners", changefreq: "monthly", priority: 0.6 },
+  { path: "/privacy", changefreq: "yearly", priority: 0.3 },
+  { path: "/terms", changefreq: "yearly", priority: 0.3 },
+];
